requirements.txt
Django==2.0.3
djangorestframework==3.7.7
django-filter==1.1.0